Update to version 3.3 of the Godot library.
