+++ samples/webview_demo/src/main/AndroidManifest.xml
@@ -4,12 +4,20 @@
 
     <uses-permission android:name="android.permission.INTERNET" />
 
+    <!-- Oculus Quest hand tracking -->
+    <uses-permission android:name="com.oculus.permission.HAND_TRACKING" />
+    <uses-feature android:name="oculus.software.handtracking" android:required="false" />
+
     <uses-feature android:glEsVersion="0x00020000" android:required="true"/>
 
     <uses-feature
         android:name="android.hardware.vr.headtracking"
         android:required="false"
         android:version="1"/>
+
+    <!-- Enable overlay keyboard -->
+    <uses-feature android:name="oculus.software.overlay_keyboard"
+        android:required="true" />
 
     <application
         android:icon="@mipmap/ic_launcher"
