+++ samples/webview_demo/src/main/java/org/godotengine/plugin/gast/samples/webview/MainActivity.kt
@@ -1,20 +1,17 @@
 package org.godotengine.plugin.gast.samples.webview
 
-import org.godotengine.godot.Godot
+import org.godotengine.godot.FullScreenGodotApp
 import org.godotengine.godot.xr.XRMode
 import java.util.*
 
-class MainActivity : Godot() {
+class MainActivity : FullScreenGodotApp() {
 
     // Override the engine starting parameters to indicate we want VR mode.
-    override fun getCommandLine(): Array<String> {
-        val original = super.getCommandLine()
-        if (BuildConfig.FLAVOR == "ovr") {
-            val updated: Array<String> = Arrays.copyOf(original, original.size + 1)
-            updated[original.size] = XRMode.OVR.cmdLineArg
-            return updated
+    override fun getCommandLine(): List<String> {
+        return if (BuildConfig.FLAVOR == "ovr") {
+            Collections.singletonList(XRMode.OVR.cmdLineArg)
         } else {
-            return original
+            Collections.emptyList()
         }
     }
